LayoutTests/imported/w3c/web-platform-tests/html/semantics/selectors/pseudo-classes/focus-expected.txt

CONSOLE MESSAGE: line 35: TypeError: null is not an object (evaluating 'document.getElementById("iframe").contentDocument.getElementById("inputiframe").focus')
button1            
hello
content


Harness Error. harness_status.status = 1 , harness_status.message = TypeError: null is not an object (evaluating 'document.getElementById("iframe").contentDocument.getElementById("inputiframe").focus')
